박대현: 모든티어 매치리스트 수집 후 매치데이터 자동 저장 후 다시 처음부터 재시작

+++ services/Matches.service.js
@@ -12,35 +12,42 @@ class MatchesService {
 
     getLeagueList = async (leagueTier) => {
         if (leagueTier === undefined) leagueTier = 'challenger';
-        console.log('챌~마스터 티어 수집 시작');
+        console.log(leagueTier,'티어 수집 시작');
 
         const leagueSummonerList = await this.api.getLeagueList(leagueTier);
-
-        this.getSummoner(leagueSummonerList);
+        if (typeof leagueSummonerList === 'object') {
+            console.log("작동")
+            this.getSummoner(leagueSummonerList);
+        }
+        else {
+            console.log('API 호출실패로 대기');
+            this.getLeagueList(leagueTier)
+        }               
+        return "데이터 수집 시작"
     };
 
     getTierList = async (division, tier, page) => {
-        console.log(tier, '데이터수집 시작');
+        console.log(tier,page + ' 페이지 데이터수집 시작');
 
-        if (page === 6) {
+        if (page >= 3) {
             console.log(`=============${tier}List저장종료=============`);
 
             switch (tier) {
                 case 'DIAMOND':
-                    this.getTierList('I', 'PLATINUM', '0');
+                    this.getTierList('I', 'PLATINUM', '1');
                     break;
                 case 'PLATINUM':
-                    this.getTierList('I', 'GOLD', '0');
+                    this.getTierList('I', 'GOLD', '1');
                     break;
                 case 'GOLD':
-                    this.getTierList('I', 'SILVER', '0');
+                    this.getTierList('I', 'SILVER', '1');
                     break;
                 case 'SILVER':
-                    this.getTierList('I', 'BRONZE', '0');
+                    this.getTierList('I', 'BRONZE', '1');
                     break;
                 default:
                     console.log('============getTierList=========');
-                    this.saveMatchData('challenger');
+                    this.saveMatchData('CHALLENGER');
                     break;
             }
         } else {
@@ -52,14 +59,14 @@ class MatchesService {
 
             this.getSummoner(leagueSummonerList, page);
         }
+        return "데이터 수집 시작"
     };
 
     getSummoner = async (leagueSummonerList, page) => {
         let i = 0;
-        console.log(leagueSummonerList.entries.length);
-        if (leagueSummonerList.tier === Master)
-            leagueSummonerList.entries.length = 1000;
-        console.log(leagueSummonerList.entries.length);
+        console.log(leagueSummonerList.entries.length,"명의 소환사");
+        if(leagueSummonerList.tier === "MASTER") leagueSummonerList.entries = leagueSummonerList.entries.slice(0,300) // 7000명 너무많아서 일단 700명으로 제한
+        if(leagueSummonerList.entries.length < i ) i = 0 
         const saveMatchInterval = setInterval(async () => {
             if (i === leagueSummonerList.entries.length) {
                 console.log('=============저장종료=============');
@@ -72,40 +79,44 @@ class MatchesService {
                         this.getLeagueList('master');
                         break;
                     case 'MASTER':
-                        this.getTierList('I', 'DIAMOND', '0');
+                        this.getTierList('I', 'DIAMOND', '1');
                         break;
                     case 'DIAMOND':
-                        page = page + 2;
+                        // page = page + 2 실행시 문자열 + 2가 돼서 기대값은 1 + 2 = 3 이었지만 결과값은 12였음
+                        page = Number(page) + 2;
                         this.getTierList('I', 'DIAMOND', page);
                         break;
                     case 'PLATINUM':
-                        page = page + 2;
+                        page = Number(page) + 2;
                         this.getTierList('I', 'PLATINUM', page);
                         break;
                     case 'GOLD':
-                        page = page + 2;
+                        page = Number(page) + 2;
                         this.getTierList('I', 'GOLD', page);
                         break;
                     case 'SILVER':
-                        page = page + 2;
+                        page = Number(page) + 2;
                         this.getTierList('I', 'SILVER', page);
                         break;
                     case 'BRONZE':
-                        page = page + 2;
+                        page = Number(page) + 2;
                         this.getTierList('I', 'BRONZE', page);
                         break;
-                    // default : this.saveMatchData('challenger');
-                    // break;
                 }
             } else {
+
                 const summoner = await this.api.getSummoner(
                     leagueSummonerList,
                     i
                 );
-                if (typeof summoner === 'object') this.getMatchList(summoner);
-                else return console.log('API 호출실패로 대기');
-
-                i++;
+                if (typeof summoner === 'object') {
+                    i++;
+                    this.getMatchList(summoner);
+                    // i++;  밑에 겟 매치리스트는 호출했으나 i++가 안되고 계속 같은소환사를 호출했음 잘되다가 ?? 왜?? 
+                }
+                else {
+                    return console.log('API 호출실패로 대기');
+                }               
             }
         }, 1300);
     };
@@ -118,8 +129,8 @@ class MatchesService {
                 const findMatch = await this.matchesRepository.findMatchList(
                     match.matchId
                 );
-                if (!findMatch) this.matchesRepository.saveMatchList(match);
-                else console.log(`이미 저장된 매치 ${match.matchId} 입니다`);
+                if (!findMatch) return this.matchesRepository.saveMatchList(match);
+                else return 
             });
         } else {
             return console.log('API 호출실패로 대기');
@@ -127,27 +138,16 @@ class MatchesService {
     };
 
     saveMatchData = async (tier) => {
-        if (tier === undefined) tier = 'CHALLENGER';
-        let i = 22866;
-        const matchList = await this.matchesRepository.findMatch(tier);
+        if (tier === undefined) tier = 'master';
+        let i = 0;
+        const matchList = await this.matchesRepository.getMatchDataList(tier);
         console.log(matchList.length, '매치리스트');
 
         const saveMatchInterval = setInterval(async () => {
             if (i === matchList.length) {
                 console.log('=============매치저장종료=============');
                 clearInterval(saveMatchInterval);
-
-                switch (tier) {
-                    case 'CHALLENGER':
-                        this.saveMatchData('GRANDMASTER');
-                        break;
-                    case 'GRANDMASTER':
-                        this.saveMatchData('MASTER');
-                        break;
-                    default:
-                        console.log('매치 저장 종료 ');
-                }
-            } else {
+            } else {           // findOrCreate
                 const findMatchId = await this.matchesRepository.findMatchById(
                     matchList[i].matchId
                 );
@@ -157,25 +157,16 @@ class MatchesService {
                         matchList[i],
                         i
                     );
-                    switch (matchData) {
-                        case 'Request failed with status code 404':
-                            await this.matchesRepository.deleteMatchList(
-                                matchList[i].matchId
-                            );
-                            console.log('Data not found');
+                    switch (typeof matchData) {
+                        case "string" :
+                            if(matchData === 'Request failed with status code 404' ) {
+                                await this.matchesRepository.deleteMatchList(
+                                    matchList[i].matchId
+                                );
+                                console.log(matchData);    
+                            }
+                            i++
                             break;
-                        case 'Request failed with status code 429':
-                            console.log('Rate limit exceeded');
-                            return;
-                        case 'Request failed with status code 500':
-                            console.log('Internal server error');
-                            return;
-                        case 'connect ETIMEDOUT 103.240.225.13:443':
-                            console.log('API 호출실패로 대기');
-                            return;
-                        case 'Request failed with status code 503':
-                            console.log('Service unavailable');
-                            return;
                         default:
                             matchData.info.participants.map((data) => {
                                 const itemList = [
@@ -213,23 +204,36 @@ class MatchesService {
                             break;
                     }
                 } else {
-                    return;
-                    // console.log(
-                    //     `${tier}티어 ${i}번쨰와 동일한 매치데이터 존재함`
-                    // );
-                    // console.log(`${tier} 매치데이터 업데이트를 완료하였습니다`);
-                    // clearInterval(saveMatchInterval);
+                    console.log(`${tier}티어 ${i}번쨰와 동일한 매치데이터 존재함`);                         
+                    console.log(`${tier} 매치데이터 업데이트를 완료하였습니다`);
+                    clearInterval(saveMatchInterval);       
+                    switch (tier) {
+                        case 'CHALLENGER':
+                            this.saveMatchData('GRANDMASTER');
+                            break;
+                        case 'GRANDMASTER':
+                            this.saveMatchData('MASTER');
+                            break;
+                        case 'MASTER':
+                            this.saveMatchData('DIAMOND');
+                            break;
+                        case 'DIAMOND':
+                            this.saveMatchData('PLATINUM');
+                            break;
+                        case 'PLATINUM':
+                            this.saveMatchData('GOLD');
+                            break;
+                        case 'GOLD':
+                            this.saveMatchData('SILVER');
+                            break;
+                        case 'SILVER':
+                            this.saveMatchData('BRONZE');
+                            break;
+                        default:
+                            console.log('매치 저장 종료');
+                            this.getLeagueList("challenger")
 
-                    // switch (tier) {
-                    //     case 'CHALLENGER':
-                    //         this.saveMatchData('GRANDMASTER');
-                    //         break;
-                    //     case 'GRANDMASTER':
-                    //         this.saveMatchData('MASTER');
-                    //         break;
-                    //     default:
-                    //         console.log('매치 저장 종료 ');
-                    // }
+                    }
                 }
                 i++;
             }
@@ -242,6 +246,21 @@ class MatchesService {
     }
 
     getChampion = async (championId) => {
+
+        const itemList = [
+            '3001','3006','3009','3011','3020','3026','3031','3033','3036','3040',
+            '3042','3046','3047','3050','3053','3065','3071','3072','3074','3075',
+            '3078','3083','3085','3089','3091','3094','3095','3100','3102','3107',
+            '3109','3110','3111','3115','3116','3117','3119','3121','3124','3135',
+            '3139','3142','3143','3152','3153','3156','3157','3158','3165','3179',
+            '3181','3190','3193','3222','3504','3508','3742','3743','3814','3853',
+            '3857','3860','3864','4005','4401','4628','4629','4633','4636','4637',
+            '4638','4643','4644','4645','6035','6333','6609','6616','6630','6631',
+            '6632','6653','6655','6656','6662','6664','6671','6672','6673','6675',
+            '6676','6691','6692','6693','6694','6695','6696','8001','8020',
+        ];
+
+      
         const tier = [
             'CHALLENGER',
             'GRANDMASTER',
@@ -251,108 +270,6 @@ class MatchesService {
             'GOLD',
             'SILVER',
             'BRONZE',
-        ];
-
-        const itemList = [
-            '3001',
-            '3006',
-            '3009',
-            '3011',
-            '3020',
-            '3026',
-            '3031',
-            '3033',
-            '3036',
-            '3040',
-            '3042',
-            '3046',
-            '3047',
-            '3050',
-            '3053',
-            '3065',
-            '3071',
-            '3072',
-            '3074',
-            '3075',
-            '3078',
-            '3083',
-            '3085',
-            '3089',
-            '3091',
-            '3094',
-            '3095',
-            '3100',
-            '3102',
-            '3107',
-            '3109',
-            '3110',
-            '3111',
-            '3115',
-            '3116',
-            '3117',
-            '3119',
-            '3121',
-            '3124',
-            '3135',
-            '3139',
-            '3142',
-            '3143',
-            '3152',
-            '3153',
-            '3156',
-            '3157',
-            '3158',
-            '3165',
-            '3179',
-            '3181',
-            '3190',
-            '3193',
-            '3222',
-            '3504',
-            '3508',
-            '3742',
-            '3743',
-            '3814',
-            '3853',
-            '3857',
-            '3860',
-            '3864',
-            '4005',
-            '4401',
-            '4628',
-            '4629',
-            '4633',
-            '4636',
-            '4637',
-            '4638',
-            '4643',
-            '4644',
-            '4645',
-            '6035',
-            '6333',
-            '6609',
-            '6616',
-            '6630',
-            '6631',
-            '6632',
-            '6653',
-            '6655',
-            '6656',
-            '6662',
-            '6664',
-            '6671',
-            '6672',
-            '6673',
-            '6675',
-            '6676',
-            '6691',
-            '6692',
-            '6693',
-            '6694',
-            '6695',
-            '6696',
-            '8001',
-            '8020',
         ];
 
         let winRateByItemtoArray = [];
@@ -415,7 +332,7 @@ class MatchesService {
                 winRateByItemtoArray.sort((a, b) => b.pick - a.pick);
             }
         }
-        // console.log(winRateByItemtoArray)
+
         return winRateByItemtoArray;
     };
